dziala plynnie, bez wywalania bledu jak na razie, dokonczyc uchwyty na dao w adapterach

+++ app/src/main/java/com/kampoz/sketchat/activity/GroupsAndSubjectsActivity.java
@@ -73,7 +73,7 @@ public class GroupsAndSubjectsActivity extends AppCompatActivity implements
         }
 
         if (savedInstanceState != null) {
-            mCurrentGroupId = savedInstanceState.getInt("GROUP_ID");
+            mCurrentGroupId = savedInstanceState.getLong("GROUP_ID");
             if (mCurrentGroupId != 0) {
                 onGroupItemSelected(mCurrentGroupId);
             }
